refactor: order for custom hooks
